Replace abstract tests with parameterized tests
- Change the implementation of AbstractQueryTestBase and
AbstractQueryExtendedTestBase from abstract test classes to test
classes that run with the Parameterized class.
- Rename AbstractQueryTestBase to QueryCoveringIndexesTest.
- Rename AbstractQueryExtendedTestBase to
QueryWithoutCoveringIndexesTest.
- Rename AbstractQueryTestSetup to AbstractQueryTestBase
- Delete the query executor matcher, SQL only, and standard test classes
since they are now replaced by the two new query parameterized test
classes.
- Update QuerySortTest.
- Update QUeryFilterFieldsTest.
- Fix database closure issue by removing unnecessary test for blank
index name in IndexManagerTest.deleteFailOnInvalidIndexName.
